add NftCollectionView
